move utils module into app
