+++ CHANGELOG.md
@@ -27,7 +27,6 @@ The following modules are completed and published to NPM and ready to be consume
 | [Performance Monitoring](/packages/perf)                 |                [![badge](https://img.shields.io/npm/dm/@react-native-firebase/perf.svg?style=for-the-badge&logo=npm)](https://www.npmjs.com/package/@react-native-firebase/perf)                |                [![badge](https://api.rnfirebase.io/coverage/perf/badge)](https://api.rnfirebase.io/coverage/perf/detail)                |
 | [Realtime Database](/packages/database)                  |            [![badge](https://img.shields.io/npm/dm/@react-native-firebase/database.svg?style=for-the-badge&logo=npm)](https://www.npmjs.com/package/@react-native-firebase/database)            |            [![badge](https://api.rnfirebase.io/coverage/database/badge)](https://api.rnfirebase.io/coverage/database/detail)            |
 | [Remote Config](/packages/remote-config)                 |       [![badge](https://img.shields.io/npm/dm/@react-native-firebase/remote-config.svg?style=for-the-badge&logo=npm)](https://www.npmjs.com/package/@react-native-firebase/remote-config)       |       [![badge](https://api.rnfirebase.io/coverage/remote-config/badge)](https://api.rnfirebase.io/coverage/remote-config/detail)       |
-| [Utils](/packages/utils)                                 |               [![badge](https://img.shields.io/npm/dm/@react-native-firebase/utils.svg?style=for-the-badge&logo=npm)](https://www.npmjs.com/package/@react-native-firebase/utils)               |               [![badge](https://api.rnfirebase.io/coverage/utils/badge)](https://api.rnfirebase.io/coverage/utils/detail)               |
 
 ---
 
@@ -144,7 +143,7 @@ AdMob has undergone a full rewrite to keep up-to-date with the latest changes an
 - [NEW] Added support for custom network extras on ad requests via `networkExtras`.
   - The user reward is now pre-fetched when the ad is loaded.
 - [BREAKING] The API interface for interacting with AdMob has undergone a full re-write.
-- [BUGFIX] Ads can now work during React Native debugging. 
+- [BUGFIX] Ads can now work during React Native debugging.
 
 ---
