+++ docs/index.md
@@ -45,4 +45,3 @@ for installation instructions on each individual module page.
 | <Anchor href="/perf">Performance Monitoring</Anchor>       | <Badge type="downloads">@react-native-firebase/perf</Badge>                | <Badge type="coverage">perf</Badge>                |
 | <Anchor href="/database">Realtime Database</Anchor>        | <Badge type="downloads">@react-native-firebase/database</Badge>            | <Badge type="coverage">database</Badge>            |
 | <Anchor href="/config">Remote Config</Anchor>              | <Badge type="downloads">@react-native-firebase/remote-config</Badge>       | <Badge type="coverage">remote-config</Badge>       |
-| <Anchor href="/utils">Utils</Anchor>                       | <Badge type="downloads">@react-native-firebase/utils</Badge>               | <Badge type="coverage">utils</Badge>               |
