+++ docs/utils/quick-start.md
@@ -10,7 +10,7 @@ description: Getting started with the Utils package in React Native Firebase
 Install this module with Yarn:
 
 ```bash
-yarn add @react-native-firebase/utils
+yarn add @react-native-firebase/app
 ```
 
 > Integrating manually and not via React Native auto-linking? Check the setup instructions for <Anchor version group href="/android">Android</Anchor> & <Anchor version group href="/ios">iOS</Anchor>.
@@ -20,13 +20,13 @@ yarn add @react-native-firebase/utils
 Import the Performance Monitoring package into your project:
 
 ```js
-import utils from '@react-native-firebase/utils';
+import utils from '@react-native-firebase/app/lib/utils';
 ```
 
 The package also provides access to the firebase instance:
 
 ```js
-import { firebase } from '@react-native-firebase/utils';
+import { firebase } from '@react-native-firebase/app/lib/utils';
 ```
 
 ### Detect whether the app is running within TestL Lab
@@ -40,7 +40,7 @@ It is useful to change the apps configuration if it is being run in Test Lab, fo
 data collection. Such functionality can be carried out by taking advantage of the `isRunningInTestLab` property:
 
 ```js
-import utils from '@react-native-firebase/utils';
+import utils from '@react-native-firebase/app/lib/utils';
 import analytics from '@react-native-firebase/analytics';
 
 async function bootstrap() {
