+++ packages/app/android/src/reactnative/java/io/invertase/firebase/util/ReactNativeFirebaseUtilsModule.java
