+++ packages/app/android/src/reactnative/java/io/invertase/firebase/util/ReactNativeFirebaseUtilsPackage.java
