+++ packages/app/ios/RNFBApp/RNFBUtilsModule.m
@@ -26,11 +26,11 @@
 #pragma mark -
 #pragma mark Module Setup
 
-  RCT_EXPORT_MODULE();
+RCT_EXPORT_MODULE();
 
-  - (dispatch_queue_t)methodQueue {
+- (dispatch_queue_t)methodQueue {
     return dispatch_get_main_queue();
-  }
+}
 
 #pragma mark -
 #pragma mark Firebase Utils Methods
