+++ packages/app/lib/utils/UtilsStatics.js
