--- packages/utils/.npmignore
@@ -1,66 +1,0 @@
-# Built application files
-android/*/build/
-
-# Crashlytics configuations
-android/com_crashlytics_export_strings.xml
-
-# Local configuration file (sdk path, etc)
-android/local.properties
-
-# Gradle generated files
-android/.gradle/
-
-# Signing files
-android/.signing/
-
-# User-specific configurations
-android/.idea/gradle.xml
-android/.idea/libraries/
-android/.idea/workspace.xml
-android/.idea/tasks.xml
-android/.idea/.name
-android/.idea/compiler.xml
-android/.idea/copyright/profiles_settings.xml
-android/.idea/encodings.xml
-android/.idea/misc.xml
-android/.idea/modules.xml
-android/.idea/scopes/scope_settings.xml
-android/.idea/vcs.xml
-android/*.iml
-
-# Xcode
-*.pbxuser
-*.mode1v3
-*.mode2v3
-*.perspectivev3
-*.xcuserstate
-ios/Pods
-ios/build
-*project.xcworkspace*
-*xcuserdata*
-
-# OS-specific files
-.DS_Store
-.DS_Store?
-._*
-.Spotlight-V100
-.Trashes
-ehthumbs.db
-Thumbs.dbandroid/gradle
-android/gradlew
-android/build
-android/gradlew.bat
-android/gradle/
-
-.idea
-coverage
-yarn.lock
-e2e/
-.github
-.vscode
-.nyc_output
-android/.settings
-*.coverage.json
-.circleci
-.eslintignore
-type-test.ts
