--- packages/utils/README.md
@@ -1,62 +1,0 @@
-<p align="center">
-  <a href="https://invertase.io/oss/react-native-firebase">
-    <img width="160px" src="https://i.imgur.com/JIyBtKW.png"><br/>
-  </a>
-  <h2 align="center">React Native Firebase - Utils</h2>
-</p>
-
-<p align="center">
-  <a href="https://api.rnfirebase.io/coverage/utils/detail"><img src="https://api.rnfirebase.io/coverage/utils/badge?style=flat-square" alt="Coverage"></a>
-  <a href="https://www.npmjs.com/package/@react-native-firebase/utils"><img src="https://img.shields.io/npm/dm/@react-native-firebase/utils.svg?style=flat-square" alt="NPM downloads"></a>
-  <a href="https://www.npmjs.com/package/@react-native-firebase/utils"><img src="https://img.shields.io/npm/v/@react-native-firebase/utils.svg?style=flat-square" alt="NPM version"></a>
-  <a href="/LICENSE"><img src="https://img.shields.io/npm/l/react-native-firebase.svg?style=flat-square" alt="License"></a>
-  <a href="https://lerna.js.org/"><img src="https://img.shields.io/badge/maintained%20with-lerna-cc00ff.svg?style=flat-square" alt="Maintained with Lerna"></a>
-</p>
-
-<p align="center">
-  <a href="https://invertase.link/discord"><img src="https://img.shields.io/discord/295953187817521152.svg?style=flat-square&colorA=7289da&label=Chat%20on%20Discord" alt="Chat on Discord"></a>
-  <a href="https://twitter.com/rnfirebase"><img src="https://img.shields.io/twitter/follow/rnfirebase.svg?style=flat-square&colorA=1da1f2&colorB=&label=Follow%20on%20Twitter" alt="Follow on Twitter"></a>
-</p>
-
----
-
-> This is for the upcoming v6.0.0 release of React Native Firebase, please use the [react-native-firebase](https://www.npmjs.com/package/react-native-firebase) package instead, unless you're early adopting/testing the new packages - in which case please use the `latest` tagged patch-only release for this package.
-
-A collection of utilities to aid in using Firebase and related services inside React Native, e.g. Test Lab helpers and Google Play Services version helpers.
-
-[> Learn More](https://firebase.google.com/products/utils/)
-
-## Installation
-
-Requires `@react-native-firebase/app` to be installed.
-
-```bash
-yarn add @react-native-firebase/utils
-react-native link @react-native-firebase/utils
-```
-
-## Documentation
-
-- [Guides](https://invertase.io/oss/react-native-firebase/guides?tags=utils)
-- [Installation](https://invertase.io/oss/react-native-firebase/v6/utils)
-- [Reference](https://invertase.io/oss/react-native-firebase/v6/utils/reference)
-
-## License
-
-- See [LICENSE](/LICENSE)
-
----
-
-<p>
-  <img align="left" width="75px" src="https://static.invertase.io/assets/invertase-logo-small.png"> 
-  <p align="left">  
-    Built and maintained with 💛 by <a href="https://invertase.io">Invertase</a>.
-  </p>
-  <p align="left">  
-    <a href="https://invertase.io/hire-us">💼 Hire Us</a> | 
-    <a href="https://opencollective.com/react-native-firebase">☕️ Sponsor Us</a> | 
-    <a href="https://opencollective.com/jobs">‍💻 Work With Us</a>
-  </p>
-</p>
-
----
