--- packages/utils/android/.editorconfig
@@ -1,10 +1,0 @@
-# editorconfig
-root = true
-
-[*]
-indent_style = space
-indent_size = 2
-end_of_line = lf
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
