--- packages/utils/android/build.gradle
@@ -1,61 +1,0 @@
-buildscript {
-  repositories {
-    google()
-    jcenter()
-  }
-  dependencies {
-    classpath 'com.android.tools.build:gradle:3.4.2'
-  }
-}
-
-plugins {
-  id "io.invertase.gradle.build" version "1.3"
-}
-
-project.ext {
-  set('react-native', [
-    versions: [
-      android           : [
-        minSdk    : 16,
-        targetSdk : 28,
-        compileSdk: 28,
-        // optional as gradle.buildTools comes with one by default
-        // overriding here though to match the version RN uses
-        buildTools: "28.0.3"
-      ],
-
-      googlePlayServices: [
-        base: "16.1.0",
-      ],
-    ],
-  ])
-}
-
-android {
-  defaultConfig {
-    multiDexEnabled true
-  }
-  lintOptions {
-    disable 'GradleCompatible'
-    abortOnError false
-  }
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_8
-    targetCompatibility JavaVersion.VERSION_1_8
-  }
-}
-
-repositories {
-  google()
-  jcenter()
-}
-
-dependencies {
-  api project(':@react-native-firebase_app')
-  implementation "com.google.android.gms:play-services-base:${ReactNative.ext.getVersion("googlePlayServices", "base")}"
-}
-
-ReactNative.shared.applyPackageVersion()
-ReactNative.shared.applyDefaultExcludes()
-ReactNative.module.applyAndroidVersions()
-ReactNative.module.applyReactNativeDependency("api")
