--- packages/utils/android/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Tue Oct 09 01:55:27 BST 2018
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-all.zip
