--- packages/utils/android/lint.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<lint>
-  <!-- Disable the given check in this project -->
-  <issue id="GradleCompatible" severity="ignore" />
-</lint>
