--- packages/utils/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = '@react-native-firebase_utils'
