--- packages/utils/e2e/utils.e2e.js
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2016-present Invertase Limited & Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this library except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-describe('utils()', () => {
-  describe('namespace', () => {
-    it('accessible from firebase.app()', () => {
-      const app = firebase.app();
-      should.exist(app.utils);
-      app.utils().app.should.equal(app);
-    });
-  });
-
-  describe('isRunningInTestLab', () => {
-    it('returns true or false', () => {
-      should.equal(firebase.utils().isRunningInTestLab, false);
-    });
-  });
-});
