--- packages/utils/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@react-native-firebase/utils",
-  "version": "0.2.3",
-  "author": "Invertase <oss@invertase.io> (http://invertase.io)",
-  "description": "React Native Firebase - Utils",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "scripts": {
-    "build": "genversion --semi lib/version.js",
-    "build:clean": "rimraf android/build && rimraf ios/build",
-    "prepare": "yarn run build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/invertase/react-native-firebase/tree/master/packages/utils"
-  },
-  "license": "Apache-2.0",
-  "keywords": [
-    "react",
-    "react-native",
-    "firebase",
-    "test lab",
-    "google play services",
-    "android",
-    "ios",
-    "utils"
-  ],
-  "peerDependencies": {
-    "@react-native-firebase/app": "*"
-  },
-  "dependencies": {
-    "@react-native-firebase/app-types": "0.2.3"
-  },
-  "gitHead": "889bf5379c8ce5257b33fb3ad1ac9c905fe26fff",
-  "publishConfig": {
-    "access": "public"
-  }
-}
